requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
avro==1.12.0
fastavro==1.9.7
argon2-cffi-bindings==25.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
confluent-kafka==2.11.1
dask==2025.9.1
dnspython==2.8.0
fastapi==0.118.0
fsspec==2025.9.0
h11==0.16.0
hudi==0.4.0
idna==3.10
importlib_metadata==8.7.0
locket==1.0.0
lz4==4.4.4
minio==7.2.18
mongomock==4.3.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.3
partd==1.4.2
psycopg2==2.9.10
py4j==0.10.9.7
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pymongo==4.15.2
pyodbc==5.2.0
pyspark==3.5.1
pytest-mongodb==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
sentinels==1.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
toolz==1.0.0
trino==0.336.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
zipp==3.23.0
zstandard==0.25.0
# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
# Job Management
celery==5.3.4
apscheduler==3.10.4
# Hudi Integration
pyhudi==0.14.0
